Update Main.sql
(a) Instructions 
The code needs to be copied and pasted on MySQL workbench. You can run it all at once. While running, you will notice, two errors are being populated in the messages. Please note that, it is intended to work like that since two of our triggers have been programmed to throw error messages. 

(b) The details can be found on the attached files - Report & PowerPoint presentation.